requirements.txt
Django==3.2
django-tagging==0.5.0
django-taggit==1.4.0
Pillow==8.2.0
